fix: component override issues
