+++ examples/SampleApp/src/screens/ChannelListScreen.tsx
@@ -10,9 +10,8 @@ import {
   View,
 } from 'react-native';
 import { useNavigation, useScrollToTop } from '@react-navigation/native';
-import { ChannelList, useTheme, useStableCallback, ChannelActionItem, WithComponents } from 'stream-chat-react-native';
+import { ChannelList, useTheme, useStableCallback, ChannelActionItem } from 'stream-chat-react-native';
 import { Channel } from 'stream-chat';
-import { ChannelPreview } from '../components/ChannelPreview';
 import { ChatScreenHeader } from '../components/ChatScreenHeader';
 import { MessageSearchList } from '../components/MessageSearch/MessageSearchList';
 import { useAppContext } from '../context/AppContext';
@@ -74,8 +73,6 @@ const options = {
   watch: true,
   message_limit: 25,
 };
-
-const HeaderNetworkDownIndicator = () => null;
 
 export const ChannelListScreen: React.FC = () => {
   const { chatClient } = useAppContext();
@@ -249,23 +246,16 @@ export const ChannelListScreen: React.FC = () => {
         )}
         <View style={{ flex: searchQuery ? 0 : 1 }}>
           <View style={[styles.channelListContainer, { opacity: searchQuery ? 0 : 1 }]}>
-            <WithComponents
-              overrides={{
-                HeaderNetworkDownIndicator,
-                Preview: ChannelPreview,
-              }}
-            >
-              <ChannelList
-                additionalFlatListProps={additionalFlatListProps}
-                filters={filters}
-                maxUnreadCount={99}
-                onSelect={onSelect}
-                options={options}
-                setFlatListRef={setScrollRef}
-                getChannelActionItems={getChannelActionItems}
-                sort={sort}
-              />
-            </WithComponents>
+            <ChannelList
+              additionalFlatListProps={additionalFlatListProps}
+              filters={filters}
+              maxUnreadCount={99}
+              onSelect={onSelect}
+              options={options}
+              setFlatListRef={setScrollRef}
+              getChannelActionItems={getChannelActionItems}
+              sort={sort}
+            />
           </View>
         </View>
       </View>
